src/mini-sdd/skills/mini-sdd-context/evals/files/eval-3/pyproject.toml
[tool.poetry]
name = "dbsnap"
version = "0.2.1"
description = "CLI tool to snapshot and diff PostgreSQL schemas"
authors = ["Marco Rossi <marco@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
psycopg2-binary = "^2.9.9"
rich = "^13.7.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
dbsnap = "dbsnap.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
